Ensure queue is populated before reading from it
When a test dies before any TAP output is generated, like when it fails to
compile, the _queue will be empty.  In this case, if --timer is given, we
try to access an undefined value, causing TAP::Formatter::JUnit::Session
to die, which kills the TAP::Harness along with it.  Add a check to avoid
this issue, so the harness can continue testing.